+++ org.eclipse.winery.repository.ui/src/app/wineryMainModules/policyTypes/policyTypeRouter.module.ts
@@ -19,7 +19,9 @@ import { InheritanceComponent } from '../../instance/sharedComponents/inheritanc
 import { PropertiesDefinitionComponent } from '../../instance/sharedComponents/propertiesDefinition/propertiesDefinition.component';
 import { LanguageComponent } from '../../instance/policyTypes/language/language.component';
 import { AppliesToComponent } from '../../instance/policyTypes/appliesTo/appliesTo.component';
-import { TemplatesOfTypeComponent } from '../../instance/sharedComponents/templatesOfTypes/templatesOfTypes.component';
+import { WineryReadmeComponent } from '../../wineryReadmeModule/wineryReadme.component';
+import { WineryLicenseComponent } from '../../wineryLicenseModule/wineryLicense.component';
+import { ImplementationsComponent } from '../../instance/sharedComponents/implementations/implementations.component';
 
 const toscaType = ToscaTypes.PolicyType;
 
@@ -31,13 +33,15 @@ const policyTypeRoutes: Routes = [
         component: InstanceComponent,
         resolve: { resolveData: InstanceResolver },
         children: [
+            { path: 'readme', component: WineryReadmeComponent },
+            { path: 'license', component: WineryLicenseComponent},
             { path: 'language', component: LanguageComponent },
             { path: 'appliesto', component: AppliesToComponent },
             { path: 'propertiesdefinition', component: PropertiesDefinitionComponent },
             { path: 'inheritance', component: InheritanceComponent },
             { path: 'documentation', component: DocumentationComponent },
             { path: 'xml', component: EditXMLComponent },
-            { path: 'templates', component:  TemplatesOfTypeComponent},
+            { path: 'templates', component: ImplementationsComponent },
             { path: '', redirectTo: 'language', pathMatch: 'full' }
 
         ]