+++ org.eclipse.winery.repository.ui/src/app/wineryReadmeModule/wineryReadme.service.ts
@@ -8,10 +8,9 @@
  */
 import { Injectable } from '@angular/core';
 import { Headers, Http, RequestOptions } from '@angular/http';
-import { InstanceService } from '../../instance.service';
 import { Observable } from 'rxjs/Observable';
-import { backendBaseURL } from '../../../configuration';
-import { ArtifactResourceApiData } from '../artifactSource/ArtifactResourceApiData';
+import { InstanceService } from '../instance/instance.service';
+import { backendBaseURL } from '../configuration';
 
 @Injectable()
 export class ReadmeService {
@@ -23,15 +22,15 @@ export class ReadmeService {
     getData(): Observable<string> {
         const headers = new Headers({ 'Accept': 'text/plain' });
         const options = new RequestOptions({ headers: headers });
-        return this.http.get(backendBaseURL + this.sharedData.path + '/source/README.md', options)
+        return this.http.get(backendBaseURL + this.sharedData.path + '/README.md', options)
             .map(res => res.text());
     }
 
-    save(readmeFile: ArtifactResourceApiData) {
+    save(readmeFile: String) {
         const headers = new Headers({ 'Content-Type': 'application/json' });
         const options = new RequestOptions({ headers: headers });
 
-        return this.http.put(backendBaseURL + this.sharedData.path + '/source/README.md', readmeFile, options)
+        return this.http.put(backendBaseURL + this.sharedData.path + '/README.md', readmeFile, options)
             .map(res => res.json());
     }