+++ CHANGELOG.md
@@ -15,7 +15,12 @@ This is similar to [Angular's CHANGELOG.md](https://github.com/angular/angular/b
 ## [unreleased]
 
 ### Changed
+- Add possibility to work with sub directories in artifact templates
+- Add edit properties as XML
+- Fix lifecycle interface is now selected after click on "generate lifecycle interface"
+- Add available/not available indicator for LFS in git log component
 - Add ZIP-button for artifactemplate-files
+- Add license and readme support for all components
 - Fix delete dialog message text
 - Fix popup text of upload message
 - Add Git Log View to track/discard changes and create commits